Move spaceship assignment logic into class
