+++ index.html
@@ -67,8 +67,6 @@
     <script src="src/js/projects/SpaceDisposalProject.js"></script>
     <script src="src/js/projects/CargoRocketProject.js"></script>
     <script src="src/js/projectsUI.js"></script>
-    <script src="src/js/spaceship.js"></script>
-    <script src="src/js/spaceshipUI.js"></script>
     <script src="src/js/research.js"></script>
     <script src="src/js/skills.js"></script>
     <script src="src/js/skillsUI.js"></script>
